Enumeration works properly with new ops and biases
Until now it was hardcoded what order parameters it should record and
output. I also wrapped the JSON file type classes's reading bits with a
try except statement so that I can see which filetype is throwing the
exception. Also, the extension of the ordeer parameter output file is
now .ops and it's first line has the tags, not the labels, to make
processing and analysis easier.