Fixed PTMC mistake and added 2D variant.
PTMC had some redundant code that was not updated to use the stacking in
the previous bug fix. So the master node was not including the stacking
energy in the energy used for acceptance testing. I have added a 2D
variant that exchanges along temperature and stacking energy. In doing
so I have reorganized the PTMC code such that there is now an abstract
class that the 1D and 2D variants derive from.