Complete refactor of MBAR script and modules.
I have created datatypes for the output files that are intended to be
read into memory at once (in contrast to those in io.py, which are read
in one line at a time). The MBAR module contains types that are
collections of simulation results across a set of conditions, and the
set of conditions is itself its own type. I have not yet implemented any
of the analysis on the pymbar.mybar object.