Changed to cmake; move some scripts.
cmake will now automatically download jsoncpp if needed, and fill find
a local installation of boost and an mpi implementation. The cmake
script includes many warning flags for the compiler and treats them as
errors, and also runs clang-tidy, and also treats those warnings as
errors. The generated makefile will recompile when only a header is
changed.