Created class to handle overcount multiplier.
This allows the overcount calculator to be swapped out for different
cases. I currently have one for staples with a maximum of two
binding-domains a with any potential, and one for dissalowed binding
potentials (where there is no restriction on the length of the staple).
I can make a more general one that calls a growthpoint enumerator, but
it would require a fair amount of modification, and be slower than the
overcount calculators for these more specific cases.